requirements.txt
Click==7.0
cycler==0.10.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
Pillow==6.2.1
pip-tools==4.2.0
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0